# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Swagger API Document",
        "title": "API for POS",
        "version": "v1"
    },
    "schemes": "HTTP",
    "basePath": "/",
    "components": {
        "securitySchemes": {
            "basicAuth": {
                "type": "http",
                "scheme": "basic",
                "name": "X-POS-User-Token",
                "in": "header"
            }
        }
    },
...
    "paths": {
...
        "/files/{filename}": {
            "post": {
                "tags": [
                    "file API"
                  ],
                "summary": "upload a file",
                "parameters": [{
                    "name": "filename",
                    "in": "path",
                    "description": "name of upload file",
                    "required": true,
                    "type": "string"
                }, {
                    "name": "file",
                    "in": "formData",
                    "type": "file",
                    "description": "file of upload",
                    "required": true
                }],
                "responses": {
                    "200": {
                        "description": "upload success"
                    },
                    "405": {
                        "description": "invalid input"
                    }
                },
                "security": [{
                    "oAuth": []
                }]
            },
            "get": {
                "tags": [
                    "file API"
                  ],
                "summary": "download a file",
                "parameters": [{
                    "name": "filename",
                    "in": "path",
                    "description": "name of download file",
                    "required": true,
                    "type": "string"
                }],
                "responses": {
                    "200": {
                        "description": "download success"
                    }
                },
                "security": [{
                    "oAuth": []
                }]
            }
        },
        "/files": {
            "get": {
                "tags": [
                    "file API"
                  ],
                "summary": "get file list of a account",
                "responses": {
                    "200": {
                        "description": "list of files"
                    }
                },
                "security": [{
                    "oAuth": []
                }]
            }
        }       
    }
}